Move the map from index.ios.js to App.js
