+++ package.json
@@ -14,6 +14,7 @@
 	"devDependencies": {
 		"babel-jest": "21.0.2",
 		"babel-preset-react-native": "3.0.2",
+		"eslint-config-rallycoding": "^3.2.0",
 		"jest": "21.0.2",
 		"react-test-renderer": "16.0.0-alpha.12"
 	},
